+++ src/android/com/outsystems/firebase/cloudmessaging/OSFirebaseCloudMessaging.kt
@@ -1,5 +1,6 @@
 package com.outsystems.firebase.cloudmessaging;
 
+import android.content.Context
 import com.outsystems.plugins.firebasemessaging.controller.*
 import com.outsystems.plugins.firebasemessaging.model.FirebaseMessagingError
 import com.outsystems.plugins.firebasemessaging.model.database.DatabaseManager
@@ -28,6 +29,8 @@ class OSFirebaseCloudMessaging : CordovaImplementation() {
         notificationManager = FirebaseNotificationManager(getActivity(), databaseManager)
         messagingManager = FirebaseMessagingManager()
         controller = FirebaseMessagingController(controllerDelegate, messagingManager, notificationManager)
+
+        setupChannelNameAndDescription()
     }
 
     private val controllerDelegate = object: FirebaseMessagingInterface {
@@ -49,7 +52,7 @@ class OSFirebaseCloudMessaging : CordovaImplementation() {
             sendPluginResult(null, Pair(error.code, error.description))
         }
     }
-    
+
     override fun execute(action: String, args: JSONArray, callbackContext: CallbackContext): Boolean {
         this.callbackContext = callbackContext
         CoroutineScope(Default).launch {
@@ -126,4 +129,30 @@ class OSFirebaseCloudMessaging : CordovaImplementation() {
     private fun setBadgeNumber() {
         controller.setBadgeNumber()
     }
+
+    private fun setupChannelNameAndDescription(){
+        val channelName = getActivity().getString(getStringResourceId("notification_channel_name"))
+        val channelDescription = getActivity().getString(getStringResourceId("notification_channel_description"))
+
+        if(!channelName.isNullOrEmpty()){
+            val editorName = getActivity().getSharedPreferences(CHANNEL_NAME_KEY, Context.MODE_PRIVATE).edit()
+            editorName.putString(CHANNEL_NAME_KEY, channelName)
+            editorName.apply()
+        }
+        if(!channelDescription.isNullOrEmpty()){
+            val editorDescription = getActivity().getSharedPreferences(CHANNEL_DESCRIPTION_KEY, Context.MODE_PRIVATE).edit()
+            editorDescription.putString(CHANNEL_DESCRIPTION_KEY, channelDescription)
+            editorDescription.apply()
+        }
+    }
+
+    private fun getStringResourceId(typeAndName: String): Int {
+        return getActivity().resources.getIdentifier(typeAndName, "string", getActivity().packageName)
+    }
+
+    companion object {
+        private const val CHANNEL_NAME_KEY = "notification_channel_name"
+        private const val CHANNEL_DESCRIPTION_KEY = "notification_channel_description"
+    }
+
 }